Document file split
